Rename test classes to Test + ClassName
